server/requirements.txt
asgiref==3.8.1
attrs==25.3.0
Django==5.2
django-stubs==5.1.3
django-stubs-ext==5.1.3
djangorestframework==3.16.0
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.5.1
Faker==37.1.0
gunicorn==23.0.0
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
packaging==25.0
pillow==11.2.1
psycopg2-binary==2.9.10
python-dotenv==1.1.0
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.25.1
sqlparse==0.5.3
types-PyYAML==6.0.12.20250516
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
